server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.56.0
shellingham==1.5.4
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
